+++ Drivers/CMSIS/DSP/Source/StatisticsFunctions/arm_kullback_leibler_f16.c
@@ -81,7 +81,7 @@ float16_t arm_kullback_leibler_f16(const float16_t * pSrcA,const float16_t * pSr
 
     accum = 0.0f16;
 
-    f16x8_t         vSum = vdupq_n_f16(0.0f16);
+    f16x8_t         vSum = vdupq_n_f16(0.0f);
     blkCnt = blockSize >> 3;
     while(blkCnt > 0)
     {
@@ -108,7 +108,7 @@ float16_t arm_kullback_leibler_f16(const float16_t * pSrcA,const float16_t * pSr
     {
        pA = *pSrcA++;
        pB = *pSrcB++;
-       accum += pA * (_Float16)logf((float32_t)pB / (float32_t)pA);
+       accum += pA * logf(pB / pA);
        
        blkCnt--;
     
@@ -134,7 +134,7 @@ float16_t arm_kullback_leibler_f16(const float16_t * pSrcA,const float16_t * pSr
     {
        pA = *pInA++;
        pB = *pInB++;
-       accum += pA * (_Float16)logf((float32_t)pB / (float32_t)pA);
+       accum += pA * logf(pB / pA);
        
        blkCnt--;
